+++ src/freeroute/interactive/Logfile.java
@@ -58,7 +58,7 @@ public class Logfile
                 this.pending_token = curr_ob;
                 return null;
             }
-            double f = ((Double) curr_ob).doubleValue();
+            double f = (double) curr_ob;
             if (i == 0)
             {
                 x = f;
@@ -222,7 +222,7 @@ public class Logfile
             this.pending_token = curr_ob;
             return -1;
         }
-        return (((Integer) curr_ob).intValue());
+        return ((int) curr_ob);
     }
     
     /**
